Make ImpulseForm._state protected
