+++ packages/react-impulse-form/src/impulse-form-list/_impulse-form-list.ts
@@ -9,11 +9,20 @@ import { Impulse, type Scope, batch } from "../dependencies"
 import { ImpulseForm } from "../impulse-form"
 
 import type { ImpulseFormListParams } from "./_impulse-form-list-params"
-import type { ImpulseFormListState } from "./_impulse-form-list-state"
+import { ImpulseFormListState } from "./_impulse-form-list-state"
 
 export class ImpulseFormList<TElement extends ImpulseForm> extends ImpulseForm<
   ImpulseFormListParams<TElement>
 > {
+  public static _createState<TElement extends ImpulseForm>(
+    elements: ReadonlyArray<TElement>,
+  ): ImpulseFormListState<TElement> {
+    return new ImpulseFormListState<TElement>(
+      null,
+      map(elements, ImpulseForm._getState),
+    )
+  }
+
   private readonly _elements = Impulse(
     (scope) => this._state._getElements(scope),
     {
@@ -48,7 +57,8 @@ export class ImpulseFormList<TElement extends ImpulseForm> extends ImpulseForm<
           ? setter(this._elements.getValue(scope), scope)
           : setter,
 
-        ({ _state }) => {
+        (element) => {
+          const _state = ImpulseForm._getState(element)
           const child = this._state._parentOf(_state)
 
           // assign independent initial state
