+++ packages/react-impulse-form/src/impulse-form-shape/impulse-form-shape.ts
@@ -1,16 +1,8 @@
 import { isUndefined } from "~/tools/is-undefined"
-import { mapValues } from "~/tools/map-values"
-import { partitionEntries } from "~/tools/partition-entries"
 
-import { Impulse, batch } from "../dependencies"
-import { isImpulseForm } from "../impulse-form"
+import { batch } from "../dependencies"
 
-import type { ImpulseFormShape as ImpulseFormShapeImpl } from "./_impulse-form-shape"
-import {
-  ImpulseFormShapeState,
-  type ImpulseFormShapeStateFields,
-  type ImpulseFormShapeStateMeta,
-} from "./_impulse-form-shape-state"
+import { ImpulseFormShape as ImpulseFormShapeImpl } from "./_impulse-form-shape"
 import type { ImpulseFormShapeErrorSetter } from "./impulse-form-shape-error-setter"
 import type { ImpulseFormShapeFields } from "./impulse-form-shape-fields"
 import type { ImpulseFormShapeFlagSetter } from "./impulse-form-shape-flag-setter"
@@ -40,20 +32,7 @@ export function ImpulseFormShape<TFields extends ImpulseFormShapeFields>(
     error,
   }: ImpulseFormShapeOptions<TFields> = {},
 ): ImpulseFormShape<TFields> {
-  const [forms, meta] = partitionEntries(fields, isImpulseForm)
-
-  const state = new ImpulseFormShapeState(
-    null,
-
-    mapValues(
-      forms,
-      ({ _state }) => _state,
-    ) as unknown as ImpulseFormShapeStateFields<TFields>,
-
-    mapValues(meta, (field) => {
-      return Impulse(field)
-    }) as unknown as ImpulseFormShapeStateMeta<TFields>,
-  )
+  const state = ImpulseFormShapeImpl._createState(fields)
 
   batch((scope) => {
     if (!isUndefined(input)) {
